pyproject.toml
[tool.poetry]
name = "dustpan"
version = "0.1.0"
description = ""
authors = ["Dawson Booth <pypi@dawsonbooth.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/dawsonbooth/dustpan"
repository = "https://github.com/dawsonbooth/dustpan"
documentation = "https://dawsonbooth.github.io/dustpan"
keywords = []
classifiers = [
    "Environment :: Console",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
include = ["LICENSE"]
packages = [{ include = "dustpan" }]

[tool.poetry.scripts]
dustpan = "dustpan.__main__:main"

[tool.poetry.dependencies]
python = "^3.7"
toml = "^0.10.2"
colorama = "^0.4.4"
attrs = "^20.3.0"

[tool.poetry.dev-dependencies]
autoflake = "^1.4"
black = "^20.8b1"
flake8 = "^3.8.4"
ghp-import = "^1.0.1"
invoke = "^1.5.0"
isort = "^5.7.0"
mypy = "^0.790"
portray = "^1.5.2"
pyclean = "^2.0.0"
pytest = "^6.2.1"
rope = "^0.18.0"

[tool.dustpan]
include = ["test.txt"]
remove-empty-directories = true

[tool.isort]
line_length = 119
profile = "black"

[tool.black]
line-length = 119

[tool.portray]
modules = ["dustpan"]

[[tool.portray.extra_markdown_extensions]]
[tool.portray.extra_markdown_extensions.toc]
permalink = true

[[tool.portray.extra_markdown_extensions]]
[tool.portray.extra_markdown_extensions."pymdownx.highlight"]
linenums = true

[tool.portray.mkdocs]
site_name = "dustpan"
site_url = "https://dawsonbooth.github.io/dustpan/"
edit_uri = "blob/master/"

markdown_extensions = [
    "admonition",
    "codehilite",
    "extra",
    "pymdownx.details",
    "pymdownx.superfences",
]

nav = [{ Overview = "README.md" }]

[tool.portray.mkdocs.theme]
name = "material"
palette = { primary = "blue grey", accent = "red" }

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
